Definition of abstract class Container
Rename of Class Container for RootContainer. RootContainer extends
Container
Creation of class SubContainer that also extends Container. This class
contains sections/communities/filter results  of the RootContainer.

The main difference between RootContainer and SubContainer is that the
former generates the visualElements contained within it, whereas the
latter extracts the visual elements from a parametrized RootContainer.