This experiment has tremendous improvements in performance.
The strategy is to create a bottom layer with a persistent image of the
graph, and a top layer with invisible elements that become visible only
if user events affect them.

The bottom layer displays an image that is regenerated only when changes
in the graph occur. That guarantees that there is no processing
generating the image every frame but only when the image contents
change.  

The invisible elements of the top layer match the position of their
visible counterparts in the image. The invisible elements are constantly
sensing mouse events. When a mouseOver event is detected by any element,
then only that element becomes visible on the top layer. In that way the
user perceives a response to his/her action. In case of mouseClicks,
drags or any other action transforming the graph, then the bottomLayer
is rendered again.